requirements.txt
Django==3.0.7
djangorestframework==3.11.0
requests==2.24.0
django-filter==2.3.0
